Created hook for fetching pokemon pages
