+++ src/pages/index.tsx
@@ -1,45 +1,30 @@
 import type { NextPage } from "next";
 import { useState } from "react";
-import { useQuery } from "react-query";
 import Head from "next/head";
 
 import CardGrid from "../components/CardGrid";
-
-type FetchProps = {
-  offSet: number;
-  page: number | string;
-};
+import { useFetchPageHook } from "../hooks/index";
 
 const Home: NextPage = () => {
   const [pageCount, setPageCount] = useState(0);
 
+  const { data, isLoading } = useFetchPageHook(pageCount);
+
   const next = () => {
+    const { next } = data;
+    if (!next) {
+      return null;
+    }
     setPageCount((current) => current + 1);
   };
 
   const prev = () => {
+    const { previous } = data;
+    if (!previous) {
+      return null;
+    }
     setPageCount((current) => current - 1);
   };
-
-  const fetchPage = async ({ offSet, page }: FetchProps) => {
-    const response = await fetch(
-      `https://pokeapi.co/api/v2/pokemon?offset=${
-        Number(page) * offSet
-      }&limit=${offSet}`
-    );
-    if (!response.ok) {
-      throw new Error("Network response was not ok");
-    }
-    return response.json();
-  };
-
-  const { isLoading, data } = useQuery(
-    ["page", pageCount],
-    ({ queryKey }) => fetchPage({ offSet: 16, page: queryKey[1] }),
-    {
-      keepPreviousData: true,
-    }
-  );
 
   return (
     <div className="container">
